refactor(rebel): split PyRblnSyncRuntime forward decl into shared header
WarmCache.h carried an inline forward declaration of rebel's
PyRblnSyncRuntime so the V4 dispatch shim could drive the rebel runtime
from C++ without dragging in absl-dependent pyrbln_impl headers. As more
translation units start calling that API (DispatchShim, future CPU
fast-path handlers), we need one canonical declaration.

Move the forward decl to torch_rbln/csrc/rbln/rebel_runtime_decl.h. The
new header is the single source of truth for the upstream signatures we
link against — adding or changing a method touches every consumer's
include set and propagates link-time signature checks consistently. The
ABI rationale (why we forward-declare instead of vendoring the real
header) lives at the top of the new file.
